+++ docs/versionhistory.rst
@@ -4,6 +4,7 @@ Version history
 1.1.0
 -----
 
+- Added support for CBOR (Concise Binary Object Representation)
 - Added typeguard checks to fail early if arguments of wrong types are passed to functions
 
 1.0.0